task 7

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -3,7 +3,7 @@
 DELIMITER $$ ;
 CREATE PROCEDURE ComputeAverageScoreForUser(IN user_id INT)
 BEGIN
-  DECLARE average_score DECIMAL;
+  DECLARE average_score DOUBLE;
   SELECT AVG(score) INTO average_score FROM corrections WHERE corrections.user_id=user_id;
   UPDATE users SET average_score = average_score WHERE id=user_id;
 END$$
